Dev is int

+++ index.js
@@ -1,5 +1,5 @@
 const negasi = function (x, y) {
-  return parseFloat(x * y);
+  return parseInt(x * y);
 };
 
 console.log(negasi(2, 3));
